Implemented forgot password function
It now works to send a request to the backend which will create a new token to change the password. Normally, this token would be sent via email which is already implemented. But something in the config seems not to work. Therefore, we currently also log the token within the backend to copy it for testing purposes into the frontend.